Add dark theme to website
