+++ packages/website/src/elements/ButtonGray.ts
@@ -3,7 +3,7 @@ import { palette } from "styled-tools";
 
 const ButtonGray = styled(Button)`
   color: ${palette("grayscale", 1)};
-  background-color: ${palette("grayscale", -3)};
+  background-color: ${palette("background", -3)};
 `;
 
 export default ButtonGray;
