+++ packages/website/src/index.js
@@ -4,7 +4,7 @@ import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import { Provider } from "reakit";
 import "./globalStyles";
-import themeWebsite from "./theme";
+import theme from "./theme/dark";
 import getEvalInContext from "./utils/getEvalInContext";
 import parseSections from "./utils/parseSections";
 import Home from "./pages/Home";
@@ -20,7 +20,7 @@ const getInitialState = ({ allSections, ...props }) => ({
 });
 
 const App = props => (
-  <Provider initialState={getInitialState(props)} devtools theme={themeWebsite}>
+  <Provider initialState={getInitialState(props)} devtools theme={theme}>
     <Router>
       <ScrollToTop>
         <Switch>
