+++ packages/website/src/theme/components.ts
@@ -1,82 +1,8 @@
 import { css } from "reakit";
 import { prop, palette as p, ifProp } from "styled-tools";
 
-export const palette = {
-  white: "#ffffff",
-  whiteText: p("black"),
-
-  black: "#212121",
-  blackText: p("white"),
-
-  // https://coolors.co/e91e63-ec407a-f06292-f48fb1-f8bbd0
-  primary: ["#fc4577", "#fd6099", "#fd88ce", "#f48fb1", "#f8bbd0"],
-  primaryText: [p("white"), p("white"), p("black"), p("black"), p("black")],
-
-  // https://coolors.co/f44336-ef5350-e57373-ef9a9a-ffcdd2
-  danger: ["#f44336", "#ef5350", "#e57373", "#ef9a9a", "#ffcdd2"],
-  dangerText: [p("white"), p("white"), p("black"), p("black"), p("black")],
-
-  // https://coolors.co/ffdd32-ffe566-ffeb8c-fff2b2-fff8d8
-  alert: ["#ffdd32", "#ffe566", "#ffeb8c", "#fff2b2", "#fff8d8"],
-  alertText: [p("black"), p("black"), p("black"), p("black"), p("black")],
-
-  // https://coolors.co/4caf50-66bb6a-81c784-a5d6a7-c8e6c9
-  success: ["#4caf50", "#66bb6a", "#81c784", "#a5d6a7", "#c8e6c9"],
-  successText: [p("white"), p("white"), p("white"), p("black"), p("black")],
-
-  grayscale: [
-    p("black"),
-    "#414141",
-    "#616161",
-    "#9e9e9e",
-    "#bdbdbd",
-    "#eaeaea",
-    "#f5f5f5",
-    p("white")
-  ],
-  grayscaleText: [
-    p("white"),
-    p("white"),
-    p("white"),
-    p("black"),
-    p("black"),
-    p("black"),
-    p("black"),
-    p("black")
-  ],
-
-  background: [
-    p("grayscale", -4),
-    p("grayscale", -3),
-    p("grayscale", -2),
-    p("grayscale", -1)
-  ],
-  backgroundText: p("black"),
-
-  shadow: [
-    "rgba(0, 0, 0, 0.9)",
-    "rgba(0, 0, 0, 0.7)",
-    "rgba(0, 0, 0, 0.5)",
-    "rgba(0, 0, 0, 0.3)",
-    "rgba(0, 0, 0, 0.15)",
-    "rgba(0, 0, 0, 0.075)"
-  ],
-  shadowText: [
-    p("white"),
-    p("white"),
-    p("white"),
-    p("black"),
-    p("black"),
-    p("black")
-  ],
-
-  transparent: "transparent",
-  transparentText: p("black"),
-
-  border: p("shadow", -2)
-};
-
 export const Blockquote = css`
+  color: ${p("alertText", -1)};
   background-color: ${p("alert", -1)};
   border-left-color: ${p("alert", 1)};
   border-left-width: 8px;
@@ -214,7 +140,8 @@ export const List = css`
 export const Overlay = css`
   padding: 1em;
   border-radius: 0.25em;
-  border: 1px solid ${p("border")};
+  box-shadow: 0 0 0 1px ${p("shadow", -2)}, 0 4px 8px ${p("shadow", -2)},
+    0 16px 48px ${p("shadow", -2)};
 `;
 
 export const Paragraph = css`
@@ -287,23 +214,3 @@ export const TooltipArrow = css`
     fill: none;
   }
 `;
-
-export default {
-  palette,
-  Blockquote,
-  Button,
-  Code,
-  Heading,
-  Input,
-  Link,
-  List,
-  Overlay,
-  Paragraph,
-  Popover,
-  PopoverArrow,
-  Sidebar,
-  Tabs,
-  Table,
-  Tooltip,
-  TooltipArrow
-};
